typography: all fg/bgs need to be readable
All of the foreground colors against the default background and
similarly all fo the background colors against the default foreground
need to be readable combinations.

The confusion stems from the fact that ansi terminals have two extra
colors to worry about: the default foreground and default background. In
gruvbox.vim these are fg1 and bg0 respectively, but in ansi that means
we have to avoid using these colors (fg1 and bg0) in the 16-color
pallete, since the result would be invisible.

To fix it, I've substituted fg3 for fg0 and bg3 for bg0 where they had
been used in the 16-color pallete. In the original colors/gruvbox.vim
you can find that these colors are indeed used alone (with default
fore/background) so that's my precedent.